Adds test with Scenario Outline and multiple inputs

+++ CalculatorExample/Calculator.cs
@@ -2,11 +2,11 @@
 {
     public class Calculator
     {
-        public int FirstNumber { get; set; }
-        public int SecondNumber { get; set; }
+        public double FirstNumber { get; set; }
+        public double SecondNumber { get; set; }
         
 
-        public int Add()
+        public double Add()
         {
             return FirstNumber + SecondNumber;
         }
